Strip down gathering step.  Use filtering; add projection.
It now only returns the MongoDB db.collection.find() criteria for
collection, so that we can hook our filters into the same call and also
add projection on fields.

This comes with a change to all scoring components as well, in which
they need to define the fields they are interested in.